Make the contrastive loss a parent and the old implementation a child class call the old implementation isolated potential
