+++ CI/unit_tests/loss_functions/test_contrastive_loss.py
@@ -33,7 +33,7 @@ from znnl.loss_functions import ContrastiveLoss
 
 class TestContrastiveLoss:
     """
-    Class for the testing of the contrastive loss functions.
+    Class for the testing of the contrastive loss function.
     """
 
     @classmethod
@@ -41,33 +41,44 @@ class TestContrastiveLoss:
         """
         Prepare the test class
         """
-        cls.predictions = np.array([[1, 1, 2], [1, 1, 1], [0, 0, 0], [2, 1, 1]])
         cls.targets = np.array([[1, 0, 0], [0, 1, 0], [0, 0, 1], [1, 0, 0]])  # one-hot
+
+    def test_create_label_map_symmetric(self):
+        """
+        Test the creation of the label map method with symmetric interactions.
+        """
+        contrastive_loss = ContrastiveLoss()
+        sim_mask, diff_mask, idx_map = contrastive_loss.create_label_map_symmetric(
+            self.targets
+        )
+
+        # Assert shapes
+        assert sim_mask.shape == (6,)
+        assert diff_mask.shape == (6,)
+        assert idx_map[0].shape == (6,)
+        assert idx_map[1].shape == (6,)
+
+        # Assert values
+        onp.testing.assert_array_equal(sim_mask, np.array([0, 0, 1, 0, 0, 0]))
+        onp.testing.assert_array_equal(diff_mask, np.array([1, 1, 0, 1, 1, 1]))
+        onp.testing.assert_array_equal(idx_map[0], np.array([0, 0, 0, 1, 1, 2]))
+        onp.testing.assert_array_equal(idx_map[1], np.array([1, 2, 3, 2, 3, 3]))
 
     def test_create_label_map(self):
         """
         Test the creation of the label map method.
         """
         contrastive_loss = ContrastiveLoss()
-        mask_sim, mask_diff, map_idx = contrastive_loss.create_label_map(self.targets)
+        pos_mask, neg_mask = contrastive_loss.create_label_map(self.targets)
 
         # Assert shapes
-        assert mask_sim.shape == (6,)
-        assert mask_diff.shape == (6,)
-        assert map_idx[0].shape == (6,)
-        assert map_idx[1].shape == (6,)
+        assert pos_mask.shape == (4, 4)
+        assert neg_mask.shape == (4, 4)
 
         # Assert values
-        print(map_idx)
-        onp.testing.assert_array_equal(mask_sim, np.array([0, 0, 1, 0, 0, 0]))
-        onp.testing.assert_array_equal(mask_diff, np.array([1, 1, 0, 1, 1, 1]))
-        onp.testing.assert_array_equal(map_idx[0], np.array([0, 0, 0, 1, 1, 2]))
-        onp.testing.assert_array_equal(map_idx[1], np.array([1, 2, 3, 2, 3, 3]))
-
-    def test_contrastive_loss(self):
-        """
-        Test the contrastive loss call method
-        """
-        contrastive_loss = ContrastiveLoss()
-        loss = contrastive_loss(self.predictions, self.targets)
-        onp.testing.assert_almost_equal(loss, 3.9300, decimal=4)
+        onp.testing.assert_array_equal(
+            pos_mask, np.array([[0, 0, 0, 1], [0, 0, 0, 0], [0, 0, 0, 0], [1, 0, 0, 0]])
+        )
+        onp.testing.assert_array_equal(
+            neg_mask, np.array([[0, 1, 1, 0], [1, 0, 1, 1], [1, 1, 0, 1], [0, 1, 1, 0]])
+        )
